+++ feature_settings/src/main/java/com/my/axe/feature_settings/language/Language.kt
@@ -38,7 +38,8 @@ fun Language(
                     BackButton {
                         onBackPressed()
                     }
-                }
+                },
+                scrolledContainerColor = MaterialTheme.colorScheme.surface
             )
         }
     ) {