+++ feature_settings/src/main/java/com/my/axe/feature_settings/style/Appearance.kt
@@ -89,7 +89,7 @@ fun Appearance(
             BackButton {
                 onBackPressed()
             }
-        })
+        }, scrolledContainerColor = MaterialTheme.colorScheme.surface)
     }) {
         Column(
             Modifier.padding(it),